web/practice1/requirements.txt
asgiref==3.5.0
Django==3.2
django-extensions==3.1.5
django-seed==0.3.1
djangorestframework==3.13.1
Faker==13.4.0
path.py==10.6
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
seed==0.11.3
six==1.16.0
sqlparse==0.4.2
toposort==1.7
tzdata==2022.1
